Reconstruct belt direction in post-processing
